+++ be/tests/integration/conftest.py
@@ -1,3 +1,5 @@
+import functools
+
 import pytest
 import redislite
 
@@ -22,20 +24,39 @@ def app(test_redis):
     yield test_app
 
 
+def get_https(client, *args, **kwargs):
+    return client.get(
+        *args, base_url="https://localhost", follow_redirects=True, **kwargs
+    )
+
+
 @pytest.fixture
 def client(app):
     return app.test_client()
 
 
 @pytest.fixture
+def client_https(app):
+    client = app.test_client()
+
+    # monkey patch the test_client with a get method that defaults to the HTTPS scheme!
+    get_https_augment = functools.partial(get_https, client)
+    setattr(client, "get_https", get_https_augment)
+
+    return client
+
+
+@pytest.fixture
 def socketio_client_factory(app):
     class SocketIO_Client_Factory:
         def create():
+
+            client = app.test_client()
             # emulates the real socketio usage where we load index first
             # and then connect with socketio, and thus the server socketio
             # has access to our cookie!
-            client = app.test_client()
-            client.get("/")
+            client.get("/", base_url="https://localhost", follow_redirects=True)
+
             return socketio.test_client(app, flask_test_client=client)
 
     return SocketIO_Client_Factory