+++ fe/src/Challenge/ChallengePlayPage.js
@@ -1,16 +1,16 @@
 import { observer } from "mobx-react";
 import React, { useEffect } from "react";
-import RootStore from "../Store/RootStore.js";
+import { useRootStore } from "../Store/RootStore.js";
 import { ChallengeView } from "./ChallengeView.js";
 
 function get_game_id_from_url() {
   let params = new URLSearchParams(window.location.search);
-  console.log(params.keys());
   return params.get("cid");
 }
 
 const ChallengePlayPage = observer(() => {
   let cid = get_game_id_from_url();
+  const RootStore = useRootStore();
 
   useEffect(() => {
     RootStore.socket.challenge_subscribe(cid);
@@ -20,7 +20,19 @@ const ChallengePlayPage = observer(() => {
     };
   }, [cid]);
 
-  return <ChallengeView challenge={RootStore.challenges.get(cid)} />;
+  const challenge = RootStore.challenges.get(cid);
+  if (challenge == null) {
+    return <span>Loading the game!</span>;
+  }
+
+  return (
+    <ChallengeView
+      challenge={challenge}
+      move_handler={(move) =>
+        RootStore.socket.challenge_send_move({ cid: cid, move: move })
+      }
+    />
+  );
 });
 
 export default ChallengePlayPage;