+++ fe/src/Games/Amazons/AmazonsLogic.js
@@ -1,6 +1,8 @@
 import { makeObservable, observable, action } from "mobx";
 import { React } from "react";
 
+import Avatar from "@mui/material/Avatar";
+
 const initializeBoard = (config) => {
   let startingBoards = {
     "10_0": [
@@ -30,7 +32,6 @@ class AmazonsLogic {
   constructor(challenge, config) {
     makeObservable(this, {
       board: observable,
-      lastMove: observable,
       turn: observable,
       setTurn: action,
       update_game_information: action,
@@ -44,29 +45,15 @@ class AmazonsLogic {
 
   turn = null;
   board = null;
-  lastMove = [null, null, null];
 
-  // TODO move to AmazonsView
-  turn_to_color = [
+  turn_to_identity = [
     {
       name: "White",
-      badge: (
-        <img
-          style={{ width: "100%", height: "100%" }}
-          src="/images/wqueen.png"
-          alt="White Queen Piece"
-        />
-      ),
+      badge: <Avatar src="/images/wqueen.png" alt="White Queen Piece" />,
     },
     {
       name: "Black",
-      badge: (
-        <img
-          style={{ width: "100%", height: "100%" }}
-          src="/images/bqueen.png"
-          alt="Black Queen Piece"
-        />
-      ),
+      badge: <Avatar src="/images/bqueen.png" alt="Black Queen Piece" />,
     },
   ];
 
@@ -123,23 +110,6 @@ class AmazonsLogic {
     this.board = new_data.board;
     this.turn = new_data.turn;
     // XXX: what exactly am I missing here?
-  }
-
-  process_new_move(move) {
-    let from_x = move.from[0];
-    let from_y = move.from[1];
-    let to_x = move.to[0];
-    let to_y = move.to[1];
-    let shoot_x = move.shoot[0];
-    let shoot_y = move.shoot[1];
-
-    this.board[from_x][from_y] = 0;
-    this.board[to_x][to_y] = this.turn + 1;
-    this.board[shoot_x][shoot_y] = 3;
-
-    this.setTurn((this.turn + 1) % 2);
-
-    this.lastMove = [move.from, move.to, move.shoot];
   }
 }