+++ fe/src/Network/Socket.js
@@ -1,12 +1,17 @@
 import io from "socket.io-client";
-import RootStore from "../Store/RootStore.js";
 import { makeObservable, observable, action } from "mobx";
+
+export const CONNECTION_STATUS_ENUM = {
+  CONNECTING: 1,
+  ONLINE: 2,
+  OFFLINE: 3,
+};
 
 class Socket {
   active_users = 0;
   connection_status = "offline";
 
-  constructor() {
+  constructor(RootStore) {
     makeObservable(this, {
       active_users: observable,
       set_active_users: action,
@@ -15,6 +20,7 @@ class Socket {
       set_connection_status: action,
     });
     this.io = null;
+    this.RootStore = RootStore;
   }
 
   set_active_users = (new_active_users) => {
@@ -33,20 +39,20 @@ class Socket {
     this.io = io(host_is_development ? "localhost" : "strate.gg", options);
 
     this.bind_socket_listeners();
-    this.set_connection_status("connecting");
+    this.set_connection_status(CONNECTION_STATUS_ENUM.CONNECTING);
   };
 
   bind_socket_listeners = () => {
     this.io.on("connect", () => {
-      this.set_connection_status("online");
+      this.set_connection_status(CONNECTION_STATUS_ENUM.ONLINE);
     });
 
     this.io.on("disconnect", (data) => {
-      this.set_connection_status("offline");
+      this.set_connection_status(CONNECTION_STATUS_ENUM.OFFLINE);
     });
 
     this.io.on("connection-player-id", (data) => {
-      RootStore.set_client_uid(data["uid"]);
+      this.RootStore.set_client_uid(data["uid"]);
     });
 
     this.io.on("connection-info-update", (data) => {
@@ -55,7 +61,7 @@ class Socket {
 
     this.io.on("challenge-update", (data) => {
       if ("result" in data && data.result === "success") {
-        RootStore.update_challenge_information(data.challenge);
+        this.RootStore.update_challenge_information(data.challenge);
       }
     });
   };
@@ -63,9 +69,8 @@ class Socket {
   create_new_challenge = (payload) => {
     this.io.emit("challenge-create", payload, (data) => {
       if (data.result && data.result === "success") {
-        window.location.replace(
-          "/play/" + data.challenge.game_name + "?cid=" + data.challenge.cid
-        );
+        window.location.href =
+          "/play/" + data.challenge.game_name + "?cid=" + data.challenge.cid;
       }
     });
   };
@@ -73,7 +78,7 @@ class Socket {
   join_challenge = (cid) => {
     this.io.emit("challenge-join", { cid: cid }, (data) => {
       if ("result" in data && data.result === "success") {
-        RootStore.update_challenge_information(data.challenge);
+        this.RootStore.update_challenge_information(data.challenge);
       }
     });
   };